+++ src/components/Shared.test.ts
@@ -6,7 +6,7 @@ import { shared } from "./Shared";
 test("Shared.test", () => {
   const [d, dG] = diagram();
   const [os, og] = of<number>(1);
-  const [p, pp] = shared(os);
+  const [p, pDestroy, pool] = shared(os);
 
   p((v) => {
     dG(`g1_${v}`);
@@ -23,6 +23,6 @@ test("Shared.test", () => {
 
   expect(d()).toBe("g1_1|g2_1|g1_2|g2_2|g1_3|g2_3|g1_4|g2_4");
 
-  pp.destroy();
-  expect(pp.size()).toBe(0);
+  pDestroy();
+  expect(pool.size()).toBe(0);
 });